Created a new theme provider to handle all things theme.
