+++ Moda.Web/src/moda.web.reactclient/app/components/common/app-header.tsx
@@ -2,16 +2,16 @@ import { Typography } from "antd";
 import { Header } from "antd/es/layout/layout";
 import Profile from "../Profile";
 import { useContext } from "react";
-import { ThemeContext } from "@/app/layout";
+import { ThemeContext } from "@/app/components/contexts/theme-context";
 
 const { Title } = Typography;
 
 export default function AppHeader() {
-    const [currentThemeName, _] = useContext(ThemeContext)
+    const [currentThemeName, setCurrentThemeName, appBarColor, agGridTheme] = useContext(ThemeContext)
     return (
         <Header style={{
             height: 50, display: 'flex', justifyContent: 'space-between', alignItems: 'center',
-            backgroundColor: currentThemeName === 'dark' ? '#262a2c' : '#2196f3'
+            backgroundColor: appBarColor
         }}>
             <Title level={3}>Moda</Title>
             <Profile />
